Reduce number of compiler warnings by choosing unsigned or not, carefully.  Also, improve some casting / integer size choices.
git-svn-id: https://svn.atnf.csiro.au/difx@6174 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876

+++ mpifxcorr/trunk/src/mark5bfile.cpp
@@ -74,7 +74,7 @@ Mark5BDataStream::Mark5BDataStream(const Configuration * conf, int snum, int id,
 	// But the amount of excess should be large enough to encompass all reasonable amounts of interloper data
 	// Here we choose 1/10 extra as a compromise.  Might be worth a revisit in the future.
 
-	readbuffersize = (bufferfactor/numsegments)*conf->getMaxDataBytes(streamnum)*11/10;
+	readbuffersize = static_cast<int>((bufferfactor/numsegments)*conf->getMaxDataBytes(streamnum)*11LL/10LL);
 	readbuffersize -= (readbuffersize % 8); // make it a multiple of 8 bytes
 	readbufferleftover = 0;
 	readbuffer = 0;	// to be allocated via initialize();
@@ -88,7 +88,7 @@ Mark5BDataStream::Mark5BDataStream(const Configuration * conf, int snum, int id,
 
 Mark5BDataStream::~Mark5BDataStream()
 {
-	long long bytesProcessed = (m5bstats.nValidFrame + m5bstats.nLostPacket)*10016 + m5bstats.nFillByte + m5bstats.nSkippedByte;
+	long long bytesProcessed = (m5bstats.nValidFrame + m5bstats.nLostPacket)*10016LL + m5bstats.nFillByte + m5bstats.nSkippedByte;
 	cinfo << startl << "Mark5B fixing statistics: nValidFrame=" << m5bstats.nValidFrame << " nInvalidFrame=" << m5bstats.nInvalidFrame << " nSkippedByte=" << m5bstats.nSkippedByte << " nFillByte=" << m5bstats.nFillByte << " nLostPacket=" << m5bstats.nLostPacket << " nCall=" << m5bstats.nCall << endl;
 
 	if(m5bstats.nFillByte > 3*bytesProcessed/4)
@@ -202,9 +202,9 @@ int Mark5BDataStream::calculateControlParams(int scan, int offsetsec, int offset
 
 
 	// Note here a time is needed, so we only count payloadbytes
-	long long segoffns = bufferinfo[atsegment].scanns + (long long)((1000000000.0*framesin)/framespersecond);
-	bufferinfo[atsegment].controlbuffer[bufferinfo[atsegment].numsent][1] = bufferinfo[atsegment].scanseconds + ((int)(segoffns/1000000000));
-	bufferinfo[atsegment].controlbuffer[bufferinfo[atsegment].numsent][2] = ((int)(segoffns%1000000000));
+	long long segoffns = bufferinfo[atsegment].scanns + static_cast<long long>((1000000000LL*framesin)/framespersecond);
+	bufferinfo[atsegment].controlbuffer[bufferinfo[atsegment].numsent][1] = bufferinfo[atsegment].scanseconds + static_cast<int>(segoffns/1000000000LL);
+	bufferinfo[atsegment].controlbuffer[bufferinfo[atsegment].numsent][2] = static_cast<int>(segoffns%1000000000LL);
 
 	//go back to nearest frame -- here the total number of bytes matters
 	bufferindex = atsegment*readbytes + framesin*framebytes;
@@ -213,7 +213,7 @@ int Mark5BDataStream::calculateControlParams(int scan, int offsetsec, int offset
 	if(bufferindex == bufferbytes)
 	{
 		if(bufferinfo[atsegment].scan != bufferinfo[(atsegment+1)%numdatasegments].scan ||
-		   ((bufferinfo[(atsegment+1)%numdatasegments].scanseconds - bufferinfo[atsegment].scanseconds)*1000000000 +
+		   ((bufferinfo[(atsegment+1)%numdatasegments].scanseconds - bufferinfo[atsegment].scanseconds)*1000000000LL +
 		   bufferinfo[(atsegment+1)%numdatasegments].scanns - bufferinfo[atsegment].scanns - bufferinfo[atsegment].nsinc != 0))
 		{
 			cwarn << startl << "bufferindex == bufferbytes --> Mode::INVALID_SUBINT" << endl;
@@ -252,11 +252,11 @@ void Mark5BDataStream::updateConfig(int segmentindex)
 	int fps = config->getFramesPerSecond(bufferinfo[segmentindex].configindex, streamnum);
 
 	//correct the nsinc - should be number of output frames*frame time
-	bufferinfo[segmentindex].nsinc = int(((bufferbytes/numdatasegments)/fb)*(1000000000.0/double(fps)) + 0.5);
+	bufferinfo[segmentindex].nsinc = static_cast<int>(((bufferbytes/numdatasegments)/fb)*(1000000000.0/double(fps)) + 0.5);
 
 	//take care of the case where an integral number of frames is not an integral number of blockspersend - ensure sendbytes is long enough
 	//note below, the math should produce a pure integer, but add 0.5 to make sure that the fuzziness of floats doesn't cause an off-by-one error
-	bufferinfo[segmentindex].sendbytes = int(((((double)bufferinfo[segmentindex].sendbytes)* ((double)config->getSubintNS(bufferinfo[segmentindex].configindex)))/(config->getSubintNS(bufferinfo[segmentindex].configindex) + config->getGuardNS(bufferinfo[segmentindex].configindex)) + 0.5));
+	bufferinfo[segmentindex].sendbytes = static_cast<int>((static_cast<double>(bufferinfo[segmentindex].sendbytes)*static_cast<double>(config->getSubintNS(bufferinfo[segmentindex].configindex)))/(config->getSubintNS(bufferinfo[segmentindex].configindex) + config->getGuardNS(bufferinfo[segmentindex].configindex)) + 0.5);
 }
 
 void Mark5BDataStream::initialiseFile(int configindex, int fileindex)
@@ -325,7 +325,7 @@ void Mark5BDataStream::initialiseFile(int configindex, int fileindex)
 		const int n=10016, d=10000;	// numerator and denominator of framesize/payload ratio
 
 		// set byte offset to the requested time
-		dataoffset = static_cast<long long>(jumpseconds*mark5bfilesummarygetbitrate(&fileSummary)/d*n/8 + 0.5);
+		dataoffset = static_cast<long long>(jumpseconds*static_cast<long long>(mark5bfilesummarygetbitrate(&fileSummary))/d*n/8 + 0.5);
 
 		readseconds += jumpseconds;
 	}
